# api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:9090
  description: Generated server url
paths:
  /poi:
    get:
      tags:
      - point-of-interest-controller
      operationId: getPointOfInterest
      parameters:
      - name: id
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PointOfInterestDto'
    put:
      tags:
      - point-of-interest-controller
      operationId: updatePointOfInterest
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PointOfInterestDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PointOfInterestDto'
    post:
      tags:
      - point-of-interest-controller
      operationId: createPointOfInterest
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PointOfInterestDto'
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PointOfInterestDto'
    delete:
      tags:
      - point-of-interest-controller
      operationId: deletePointOfInterest
      parameters:
      - name: id
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /poi/{type}:
    get:
      tags:
      - point-of-interest-controller
      operationId: listPointsOfInterestByType
      parameters:
      - name: type
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PointOfInterestDto'
components:
  schemas:
    PointOfInterestDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        poiType:
          type: string
        addressName:
          type: string
        city:
          type: string
        zipCode:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double